fixes bug with smooth scroll when scroll not returning correctly
there is occasions where the scroll has completed but as the offsets dont match 100% then it never exits the function. this causes the user to not be able to scroll the page any longer. by adding a rounding to the window.pageYOffset the values match and allows the function to complete